requirements.txt
boto3>=1.9.177
dnspython>=1.16.0
